scrollview added to TrackerCardViewController
