+++ Tracker/SceneDelegate.swift
@@ -16,7 +16,7 @@ final class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     ) {
         guard let scene = (scene as? UIWindowScene) else { return }
         let window = UIWindow(windowScene: scene)
-        window.rootViewController = TabBarController()
+        window.rootViewController = TrackersListViewController()
         self.window = window
         window.makeKeyAndVisible()
     }
