+++ Tracker/ViewModels/ScheduleTableViewCell.swift
@@ -1,0 +1,15 @@
+//
+//  ScheduleTableViewCell.swift
+//  Tracker
+//
+//  Created by Admin on 11/4/23.
+//
+
+import UIKit
+
+struct TrackerScheduleTableViewCellViewModel {
+    var scheduleView: UIView
+    var scheduleSwitch: UISwitch
+    var scheduleLabel: UILabel
+    var scheduleFooterView: UIView
+}
